social_media_parser.py.py
import logging
from typing import List, Dict, Any, Optional
import httpx
from bs4 import BeautifulSoup
from datetime import datetime, timezone
import re

logger = logging.getLogger(__name__)

async def fetch_recent_news_from_social_media(url: str, limit: int = 10) -> List[Dict[str, Any]]:
    """
    Placeholder logic for fetching news from a social media page.
    Note: Scraping social media platforms is often difficult due to dynamic content,
    strict API limitations, and terms of service. It's recommended to use official APIs
    if available and permitted. This implementation provides a basic example of
    how you might start to scrape a page with static content.
    """
    logger.info(f"Attempting to fetch content from social media URL: {url}")
    news_items = []

    try:
        async with httpx.AsyncClient(timeout=15) as client:
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36'
            }
            response = await client.get(url, follow_redirects=True, headers=headers)
            response.raise_for_status()

            soup = BeautifulSoup(response.text, 'lxml')

            # This is a very generic and likely to fail selector. It's just a placeholder.
            # You would need to inspect the specific social media page's HTML structure.
            post_elements = soup.find_all('div', class_=re.compile("post|article"))

            for post in post_elements[:limit]:
                # Find title, content, link, etc. inside each post element
                title = post.find('h1') or post.find('h2') or post.find('h3')
                content = post.find('p')
                link = post.find('a', href=True)

                if title and content and link:
                    news_items.append({
                        'title': title.get_text(strip=True),
                        'content': content.get_text(strip=True)[:250] + '...',
                        'source_url': link['href'],
                        'image_url': None, # Requires more specific scraping logic
                        'published_at': datetime.now(timezone.utc)
                    })

    except httpx.RequestError as e:
        logger.error(f"HTTP error for {url}: {e}")
    except Exception as e:
        logger.error(f"Error fetching from social media URL {url}: {e}", exc_info=True)
    
    return news_items

