requirements.txt
aiogram==3.9.0
aiohttp==3.9.5
psycopg[binary]==3.2.9
psycopg-pool==3.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
html5lib==1.1
fastapi==0.111.0
uvicorn==0.30.1
gTTS==2.5.1
pydantic==2.8.2
feedparser==6.0.11